hide LSP events a bit better
