+++ contribs/freight/src/main/java/org/matsim/contrib/freight/controler/LSPEventCreatorUtils.java
@@ -19,31 +19,28 @@
  *
  */
 
-package org.matsim.contrib.freight.events.eventsCreator;
+package org.matsim.contrib.freight.controler;
 
-import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.events.ActivityStartEvent;
-import org.matsim.api.core.v01.events.Event;
-import org.matsim.api.core.v01.population.Activity;
-import org.matsim.api.core.v01.population.Person;
-import org.matsim.contrib.freight.carrier.Carrier;
-import org.matsim.contrib.freight.carrier.ScheduledTour;
-import org.matsim.contrib.freight.carrier.Tour.ServiceActivity;
-import org.matsim.contrib.freight.carrier.Tour.TourElement;
-import org.matsim.contrib.freight.events.LSPServiceStartEvent;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 
-/*package-private*/  final class LSPServiceStartEventCreator implements LSPEventCreator {
+final class LSPEventCreatorUtils {
 
-	@Override
-	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour, Id<Person> driverId, int activityCounter) {
-		if( event instanceof ActivityStartEvent startEvent ){
-			if( startEvent.getActType().equals( "service" ) ) {
-				TourElement element = scheduledTour.getTour().getTourElements().get(activityCounter);
-				if( element instanceof ServiceActivity serviceActivity ) {
-					return new LSPServiceStartEvent(startEvent, carrier.getId(), driverId, serviceActivity.getService(), event.getTime(), scheduledTour.getVehicle());
-				}
-			}	
-		}
-		return null;
+	private LSPEventCreatorUtils(){
 	}
-}	
+	static Collection<LSPEventCreator> getStandardEventCreators(){
+		List<LSPEventCreator> creators = new ArrayList<>();
+//		creators.add(new LSPFreightLinkEnterEventCreator());
+//		creators.add(new LSPFreightLinkLeaveEventCreator());
+//		creators.add(new LSPFreightVehicleLeavesTrafficEventCreator());
+		creators.add(new LSPServiceEndEventCreator());
+		creators.add(new LSPServiceStartEventCreator());
+		creators.add(new LSPShipmentDeliveredEventCreator());
+		creators.add(new LSPShipmentPickedUpEventCreator());
+		creators.add(new LSPTourEndEventCreator());
+		creators.add(new LSPTourStartEventCreator());
+		return creators;
+	}
+	
+}
