+++ contribs/freight/src/main/java/org/matsim/contrib/freight/controler/LSPServiceStartEventCreator.java
@@ -19,21 +19,31 @@
  *
  */
 
-package org.matsim.contrib.freight.events.eventsCreator;
+package org.matsim.contrib.freight.controler;
 
 import org.matsim.api.core.v01.Id;
+import org.matsim.api.core.v01.events.ActivityStartEvent;
 import org.matsim.api.core.v01.events.Event;
 import org.matsim.api.core.v01.population.Activity;
-import org.matsim.api.core.v01.population.Leg;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.api.core.v01.population.Route;
 import org.matsim.contrib.freight.carrier.Carrier;
-import org.matsim.contrib.freight.carrier.CarrierService;
-import org.matsim.contrib.freight.carrier.CarrierShipment;
 import org.matsim.contrib.freight.carrier.ScheduledTour;
+import org.matsim.contrib.freight.carrier.Tour.ServiceActivity;
+import org.matsim.contrib.freight.carrier.Tour.TourElement;
+import org.matsim.contrib.freight.events.LSPServiceStartEvent;
 
-public interface LSPEventCreator {
+/*package-private*/  final class LSPServiceStartEventCreator implements LSPEventCreator {
 
-	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour, Id<Person> driverId, int activityCounter);
-	
-}
+	@Override
+	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour, Id<Person> driverId, int activityCounter) {
+		if( event instanceof ActivityStartEvent startEvent ){
+			if( startEvent.getActType().equals( "service" ) ) {
+				TourElement element = scheduledTour.getTour().getTourElements().get(activityCounter);
+				if( element instanceof ServiceActivity serviceActivity ) {
+					return new LSPServiceStartEvent(startEvent, carrier.getId(), driverId, serviceActivity.getService(), event.getTime(), scheduledTour.getVehicle());
+				}
+			}	
+		}
+		return null;
+	}
+}
