+++ contribs/freight/src/main/java/org/matsim/contrib/freight/controler/LSPShipmentPickedUpEventCreator.java
@@ -19,32 +19,30 @@
  *
  */
 
-package org.matsim.contrib.freight.events.eventsCreator;
+package org.matsim.contrib.freight.controler;
 
 import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.events.ActivityStartEvent;
+import org.matsim.api.core.v01.events.ActivityEndEvent;
 import org.matsim.api.core.v01.events.Event;
 import org.matsim.api.core.v01.population.Activity;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.FreightConstants;
 import org.matsim.contrib.freight.carrier.ScheduledTour;
-import org.matsim.contrib.freight.events.LSPTourEndEvent;
+import org.matsim.contrib.freight.carrier.Tour.Pickup;
+import org.matsim.contrib.freight.events.ShipmentPickedUpEvent;
 
-/*package-private*/ final class LSPTourEndEventCreator implements LSPEventCreator {
+/*package-private*/  final class LSPShipmentPickedUpEventCreator implements LSPEventCreator {
 
 	@Override
 	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour,
 			Id<Person> driverId, int activityCounter) {
-		if(event instanceof ActivityStartEvent) {
-			ActivityStartEvent startEvent = (ActivityStartEvent) event;
-			if(startEvent.getActType().equals(FreightConstants.END)) {
-				return new LSPTourEndEvent(carrier.getId(),  driverId, scheduledTour.getTour(), startEvent.getTime(), scheduledTour.getVehicle());
+		if(event instanceof ActivityEndEvent) {
+			if(event.getEventType().equals(FreightConstants.PICKUP)) {
+				Pickup pickup = (Pickup) activity;
+				return new ShipmentPickedUpEvent(carrier.getId(), driverId, pickup.getShipment(), event.getTime());
 			}
-		}	
+		}
 		return null;
 	}
-
-	
-
 }
