+++ contribs/freight/src/main/java/org/matsim/contrib/freight/controler/LSPTourEndEventCreator.java
@@ -19,29 +19,32 @@
  *
  */
 
-package org.matsim.contrib.freight.events.eventsCreator;
+package org.matsim.contrib.freight.controler;
 
 import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.events.ActivityEndEvent;
+import org.matsim.api.core.v01.events.ActivityStartEvent;
 import org.matsim.api.core.v01.events.Event;
 import org.matsim.api.core.v01.population.Activity;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.FreightConstants;
 import org.matsim.contrib.freight.carrier.ScheduledTour;
-import org.matsim.contrib.freight.events.LSPTourStartEvent;
+import org.matsim.contrib.freight.events.LSPTourEndEvent;
 
-/*package-private*/  final class LSPTourStartEventCreator implements LSPEventCreator {
+/*package-private*/ final class LSPTourEndEventCreator implements LSPEventCreator {
 
 	@Override
-	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour, Id<Person> driverId, int activityCounter) {
-		if((event instanceof ActivityEndEvent)) {
-			ActivityEndEvent endEvent = (ActivityEndEvent) event;
-			if(endEvent.getActType().equals(FreightConstants.START)) {
-				return new LSPTourStartEvent(carrier.getId(), driverId, scheduledTour.getTour(), event.getTime(), scheduledTour.getVehicle());
-			}	
-		}
-		return null;	
+	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour,
+			Id<Person> driverId, int activityCounter) {
+		if(event instanceof ActivityStartEvent) {
+			ActivityStartEvent startEvent = (ActivityStartEvent) event;
+			if(startEvent.getActType().equals(FreightConstants.END)) {
+				return new LSPTourEndEvent(carrier.getId(),  driverId, scheduledTour.getTour(), startEvent.getTime(), scheduledTour.getVehicle());
+			}
+		}	
+		return null;
 	}
+
+	
 
 }
