+++ contribs/freight/src/main/java/org/matsim/contrib/freight/controler/LSPTourStartEventCreator.java
@@ -19,7 +19,7 @@
  *
  */
 
-package org.matsim.contrib.freight.events.eventsCreator;
+package org.matsim.contrib.freight.controler;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.events.ActivityEndEvent;
@@ -29,20 +29,19 @@ import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.FreightConstants;
 import org.matsim.contrib.freight.carrier.ScheduledTour;
-import org.matsim.contrib.freight.carrier.Tour.Pickup;
-import org.matsim.contrib.freight.events.ShipmentPickedUpEvent;
+import org.matsim.contrib.freight.events.LSPTourStartEvent;
 
-/*package-private*/  final class LSPShipmentPickedUpEventCreator implements LSPEventCreator {
+/*package-private*/  final class LSPTourStartEventCreator implements LSPEventCreator {
 
 	@Override
-	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour,
-			Id<Person> driverId, int activityCounter) {
-		if(event instanceof ActivityEndEvent) {
-			if(event.getEventType().equals(FreightConstants.PICKUP)) {
-				Pickup pickup = (Pickup) activity;
-				return new ShipmentPickedUpEvent(carrier.getId(), driverId, pickup.getShipment(), event.getTime());
-			}
+	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour, Id<Person> driverId, int activityCounter) {
+		if((event instanceof ActivityEndEvent)) {
+			ActivityEndEvent endEvent = (ActivityEndEvent) event;
+			if(endEvent.getActType().equals(FreightConstants.START)) {
+				return new LSPTourStartEvent(carrier.getId(), driverId, scheduledTour.getTour(), event.getTime(), scheduledTour.getVehicle());
+			}	
 		}
-		return null;
+		return null;	
 	}
+
 }
